src/requirements.txt
numpy
pynput
matplotlib.pyplot
scipy
